refactor: adjust by PR
