+++ packages/mirror-tv-next/utils/common.ts
@@ -77,10 +77,22 @@ function handleApiData(apiData: string): ApiData[] {
   }
 }
 
+class FetchError extends Error {
+  public code: number
+
+  constructor(url: string, message: string = 'Not Found', code: number = 404) {
+    const errorMessage = `${message}, url: ${url}`
+    super(errorMessage)
+    this.name = this.constructor.name
+    this.code = code
+  }
+}
+
 export {
   extractYoutubeId,
   isServer,
   handleResponse,
   handleApiData,
   handleMetaDesc,
+  FetchError,
 }
